% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/jags.par.R
\name{jags.par}
\alias{jags.par}
\title{Parallel n chain jags}
\usage{
jags.par(cl, data, inits, parameters.to.save, model.file, n.iter, n.chains,
  n.thin, n.adapt = 0, n.burnin, DIC = TRUE)
}
\arguments{
\item{cl}{an object of class "cluster"}

\item{...}{other arguments. See in jags in R2jags}
}
\description{
This function parallels n.chain and runs jags.
}
\examples{
jags.par()
}
\keyword{jags.par}

